refactor(DAO-2135): move withdraw modal components into modals/withdraw/ subfolder
Move BtcWithdrawModal, WithdrawSteps, WithdrawAmountStep,
WithdrawAllowanceStep, and WithdrawReviewStep (plus tests and stories)
into modals/withdraw/. Add barrel index.ts for withdraw/ and modals/.
