+++ src/app/btc-vault/components/modals/index.ts
@@ -1,0 +1,2 @@
+export { BtcDepositModal } from './deposit/BtcDepositModal'
+export { BtcWithdrawModal } from './withdraw/BtcWithdrawModal'
