+++ src/app/btc-vault/components/modals/withdraw/BtcWithdrawModal.tsx
@@ -10,9 +10,9 @@ import { ProgressBar } from '@/components/ProgressBarNew'
 import { Header } from '@/components/Typography'
 import { VAULT_SHARE_DECIMALS } from '@/lib/constants'
 
-import { useUserPosition } from '../hooks/useUserPosition'
-import { useVaultMetrics } from '../hooks/useVaultMetrics'
-import { BTC_VAULT_WITHDRAWAL_FEE, WITHDRAWAL_STEP_PROGRESS } from '../services/constants'
+import { useUserPosition } from '../../../hooks/useUserPosition'
+import { useVaultMetrics } from '../../../hooks/useVaultMetrics'
+import { BTC_VAULT_WITHDRAWAL_FEE, WITHDRAWAL_STEP_PROGRESS } from '../../../services/constants'
 import { WithdrawAllowanceStep } from './WithdrawAllowanceStep'
 import { WithdrawAmountStep } from './WithdrawAmountStep'
 import { WithdrawReviewStep } from './WithdrawReviewStep'
