+++ src/app/btc-vault/components/modals/withdraw/WithdrawAmountStep.tsx
@@ -14,7 +14,7 @@ import { RBTC, VAULT_SHARE_DECIMALS } from '@/lib/constants'
 import { cn, formatCurrency, handleAmountInput } from '@/lib/utils'
 import { usePricesContext } from '@/shared/context'
 
-import { BTC_VAULT_WITHDRAWAL_DISCLAIMER } from '../services/constants'
+import { BTC_VAULT_WITHDRAWAL_DISCLAIMER } from '../../../services/constants'
 
 interface WithdrawAmountStepProps {
   amount: string
