+++ src/app/btc-vault/components/modals/withdraw/index.ts
@@ -1,0 +1,1 @@
+export { BtcWithdrawModal } from './BtcWithdrawModal'
